+++ src/routes/Routes.jsx
@@ -14,32 +14,29 @@ import thirdMazeData from 'test/games/maze/easy/003.json';
 import fourthMazeData from 'test/games/maze/easy/004.json';
 import MazeGameContainer from 'containers/MazeGameContainer';
 
+export type RouteDescription = {|
+  exact?: boolean,
+  main: ReactClass<*>,
+  path: string,
+  title: string,
+|};
+
 export default [{
-  id: 0,
-  path: '/',
-  exact: true,
-  title: 'Home',
-  main: () => (<h2>{'Home'}</h2>),
-}, {
-  id: 1,
   path: '/maze/first',
   exact: true,
   title: 'First Maze',
   main: () => (<MazeGameContainer blocklyData={blocklyData} gameMetadata={firstMazeData} />),
 }, {
-  id: 2,
   path: '/maze/second',
   exact: true,
   title: 'Second Maze',
   main: () => (<MazeGameContainer blocklyData={blocklyData} gameMetadata={secondMazeData} />),
 }, {
-  id: 3,
   path: '/maze/third',
   exact: true,
   title: 'Third Maze',
   main: () => (<MazeGameContainer blocklyData={blocklyData} gameMetadata={thirdMazeData} />),
 }, {
-  id: 4,
   path: '/maze/fourth',
   exact: true,
   title: 'Fourth Maze',